requirements.txt
streamlit
PyMuPDF
google-generativeai
google.ai.generativelanguage
Pillow
python-dotenv
pandas
